+++ packages/sdk-metrics/src/export/PeriodicExportingMetricReader.ts
@@ -24,6 +24,7 @@ import {
 import { MetricReader } from './MetricReader';
 import { PushMetricExporter } from './MetricExporter';
 import { callWithTimeout, TimeoutError } from '../utils';
+import { diag } from '@opentelemetry/api';
 
 export type PeriodicExportingMetricReaderOptions = {
   /**
@@ -117,11 +118,24 @@ export class PeriodicExportingMetricReader extends MetricReader {
       );
     }
 
-    const result = await internal._export(this._exporter, resourceMetrics);
-    if (result.code !== ExportResultCode.SUCCESS) {
-      throw new Error(
-        `PeriodicExportingMetricReader: metrics export failed (error ${result.error})`
-      );
+    const doExport = async () => {
+      const result = await internal._export(this._exporter, resourceMetrics);
+      if (result.code !== ExportResultCode.SUCCESS) {
+        throw new Error(
+          `PeriodicExportingMetricReader: metrics export failed (error ${result.error})`
+        );
+      }
+    };
+
+    // Avoid scheduling a promise to make the behavior more predictable and easier to test
+    if (resourceMetrics.resource.asyncAttributesPending) {
+      resourceMetrics.resource
+        .waitForAsyncAttributes?.()
+        .then(doExport, err =>
+          diag.debug('Error while resolving async portion of resource: ', err)
+        );
+    } else {
+      await doExport();
     }
   }